+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to `spring-correctness-linter` will be documented in this fi
 
 ## [Unreleased]
 
+### Added
+
+- Added lightweight external rule discovery through `ServiceLoader` via `io.github.koyan9.linter.core.spi.LintRuleProvider`
+
 ## [0.1.3] - 2026-03-27
 
 ### Added